Split dynophores.base in dynophores.core with submodules
