+++ dynophores/core/dynophore.py
@@ -1,22 +1,14 @@
 """
-Contains base functions/classes.
+dynophores.core.dynophore
+
+Handles dynophore class.
 """
 
-import math
-
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
-FEATURE_COLORS = {
-    "HBA": "firebrick",
-    "HBD": "green",
-    "H": "gold",
-    "AR": "mediumblue",
-    "PI": "blue",
-    "NI": "red",
-}
-plt.style.use("seaborn")
+from dynophores.core.superfeature import Superfeature
+from dynophores.core.envpartner import EnvPartner
 
 
 class Dynophore:
@@ -346,280 +338,3 @@ class Dynophore:
             envpartners[superfeature.id] = superfeature_envpartners
 
         return envpartners
-
-
-class Superfeature:
-    """
-    Class to store superfeature data, i.e. data on environmental partners, and important functions
-    to interact with this data.
-
-    Attributes
-    ----------
-    id : str
-        Superfeature name.
-    feature_type : str
-        Pharmacophoric feature type.
-    atom_numbers : list of int
-        List of atom IDs.
-    occurrences : pandas.DataFrame
-        Per superfeature (columns), occurrences (0=no, 1=yes) in each frame (row).
-    envpartners : list of EnvPartner
-        Superfeature's environmental partners.
-    """
-
-    def __init__(self, id, feature_type, atom_numbers, occurrences, envpartners):
-
-        self.id = id
-        self.feature_type = feature_type
-        self.atom_numbers = atom_numbers
-        self.occurrences = occurrences
-        self.envpartners = envpartners
-
-    @property
-    def envpartners_occurrences(self):
-        """
-        Get the superfeature's environmental partners' occurrences per environmental partner and
-        frame.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Occurrences (0=no, 1=yes) of an environmental partner (columns) in each frame (row).
-
-        """
-
-        return self._get_envpartners_data(type="occurrences")
-
-    @property
-    def envpartners_distances(self):
-        """
-        Get the superfeature's environmental partners' distances per environmental partner and
-        frame.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Distances to an environmental partner (columns) in each frame (row)
-        """
-
-        return self._get_envpartners_data(type="distances")
-
-    @property
-    def n_frames(self):
-        """
-        Get superfeatures's number of frames.
-
-        Returns
-        -------
-        int
-            Number of frames.
-        """
-
-        return len(self.occurrences)
-
-    @property
-    def count(self):
-        """
-        Get number of frames in which the superfeature occurs, including the superfeature's
-        environmental partners occurrences.
-
-        Returns
-        -------
-        pandas.Series
-            Superfeature count: The Series shows interactions (yes/no) to each single
-            environmental partner as well as any environmental partner.
-        """
-
-        superfeature_count = pd.Series({"any": sum(self.occurrences)})
-        envpartners_count = pd.Series(
-            {envpartner.id: envpartner.count for envpartner in self.envpartners}
-        )
-
-        return superfeature_count.append(envpartners_count)
-
-    @property
-    def frequency(self):
-        """
-        Get frequency of frames in which the superfeature occurs, including the superfeature's
-        environmental partners occurrences.
-
-        Returns
-        -------
-        pandas.Series
-            Superfeature frequency: The Series shows interactions (yes/no) to each single
-            environmental partner as well as any environmental partner.
-        """
-
-        return self.count.apply(lambda x: round(x / self.n_frames * 100, 2))
-
-    def _get_envpartners_data(self, type="occurrences"):
-        """
-        Get occurrences or distances of a superfeature's environmental partners.
-
-        Parameters
-        ----------
-        type : str
-            Data type: occurrences (default) or distances.
-
-        Returns
-        -------
-        DataFrame
-            Occurrences (default) or distances of a superfeature's environmental partners
-            (columns) for all frames (rows).
-        """
-
-        types = ["occurrences", "distances"]
-        if type in types:
-            pass
-        else:
-            raise ValueError(f'Wrong type. Select from: {", ".join(types)}')
-
-        envpartners = pd.DataFrame(
-            [getattr(envpartner, type) for envpartner in self.envpartners],
-            index=[envpartner.id for envpartner in self.envpartners],
-        ).transpose()
-
-        return envpartners
-
-
-class EnvPartner:
-    """
-    Class to store environmental partner data.
-
-    Attributes
-    ----------
-    id : str
-        ID of environmental partner.
-    residue_name : str
-        Residue name of Environmental partner.
-    residue_number : int
-        Residue number of environmental partner.
-    chain : str
-        Chain of environmental partner.
-    atom_numbers : list of int
-        List of atom numbers.
-    occurrences : numpy.array
-        Occurrences of interaction with environmental partner (0=no, 1=yes) in each frame.
-    distances : numpy.array
-        Interaction distances in each frame.
-    """
-
-    def __init__(
-        self, id, residue_name, residue_number, chain, atom_numbers, occurrences, distances
-    ):
-
-        if len(occurrences) != len(distances):
-            raise ValueError("Occurrences and distances must be of same length.")
-
-        self.id = id
-        self.residue_name = residue_name
-        self.residue_number = residue_number
-        self.chain = chain
-        self.atom_numbers = atom_numbers
-        self.occurrences = occurrences
-        self.distances = distances
-
-    @property
-    def n_frames(self):
-        """
-        Get environmental partner's number of frames.
-
-        Returns
-        -------
-        int
-            Number of frames.
-        """
-
-        return len(self.occurrences)
-
-    @property
-    def count(self):
-        """
-        Get number of frames in which the interaction with the environmental partner occurs.
-
-        Returns
-        -------
-        int
-            Count of interaction occurrence.
-        """
-
-        return sum(self.occurrences)
-
-    @property
-    def frequency(self):
-        """
-        Get frequency of frames in which the interaction with the environmental partner occurs.
-
-        Returns
-        -------
-        float
-            Frequency of interaction occurrence.
-        """
-
-        return round(sum(self.occurrences) / self.n_frames * 100, 2)
-
-
-class Plotting:
-    """
-    Class to store plotting functions
-    """
-
-    def __init__(self):
-        pass
-
-    @staticmethod
-    def plot_occurrences(occurrences, color_by_feature_type=True, max_frames=1000):
-        """
-        Create a barcode plot for superfeature or interaction occurrences.
-
-        Parameters
-        ----------
-        occurrences : pandas.DataFrame
-            Occurrences (0 or 1) per frame (rows) for one or more superfeatures or interactions
-            (columns).
-        color_by_feature_type : bool
-            Color barcode by feature type (default) or color all in black.
-        max_frames : int
-            Number of frames to display in barcode plot. If input data contains more than
-            `max_frames`, `max_frames` equidistant frames will be selected.
-        """
-        # Set plotting style
-        plt.style.use("seaborn-dark")
-
-        # Sort data by ratio
-        ratio = round(occurrences.apply(sum) / occurrences.shape[0] * 100, 2)
-        ratio.sort_values(inplace=True)
-        occurrences = occurrences[ratio.index]
-
-        # Get subset of data if more than 1000 frames
-        if occurrences.shape[0] > 1000:
-            selected_indices = [i for i in range(0, 1000, math.floor(1002 / max_frames))]
-            occurrences = occurrences.iloc[selected_indices, :]
-
-        # Transform 1 in binary values to rank in plot
-        occurrences_plot = {}
-        for i, (name, data) in enumerate(occurrences.items()):
-            data = data.replace([0, 1], [None, i + 1])
-            occurrences_plot[name] = data
-        occurrences_plot = pd.DataFrame(occurrences_plot)
-
-        # Feature type colors?
-        if color_by_feature_type:
-            feature_types = [i.split("[")[0] for i in occurrences_plot.columns]
-            colors = [
-                FEATURE_COLORS[i] if i in FEATURE_COLORS.keys() else "black" for i in feature_types
-            ]
-        else:
-            colors = None
-
-        # Plot (plot size depending on number barcodes)
-        fig, ax = plt.subplots(figsize=(10, occurrences_plot.shape[1] / 2))
-        occurrences_plot.plot(
-            marker="|", markersize=5, linestyle="", legend=None, ax=ax, color=colors
-        )
-        # Set y tick labels
-        ax.set_yticks(range(0, occurrences_plot.shape[1] + 2))
-        ax.set_yticklabels([""] + occurrences_plot.columns.to_list() + [""])
-        # Set x axis limits and label
-        ax.set_xlabel("frame index")
-        ax.set_xlim((occurrences_plot.index[0], occurrences_plot.index[-1]))
