+++ dynophores/definitions.py
@@ -1,0 +1,14 @@
+"""
+dynophores.definitions
+
+Handles definitions for dynophore library.
+"""
+
+FEATURE_COLORS = {
+    "HBA": "firebrick",
+    "HBD": "green",
+    "H": "gold",
+    "AR": "mediumblue",
+    "PI": "blue",
+    "NI": "red",
+}
